frederick-24feb

+++ Assets/Scripts/Core/PlayerHealth.cs
@@ -6,16 +6,16 @@ public class PlayerHealth : MonoBehaviour
 {
 
     [Tooltip("This is the max health of the player")]
-    public int maxHealth = 100;
+    public float maxHealth = 100;
     [Tooltip("This is the current health of the player")]
-    public int currentHealth;
+    public float currentHealth;
     // Start is called before the first frame update
     void Start()
     {
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
         if (currentHealth <= 0)
